pyproject.toml
[project]
name = "graph-rag-graph-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.8",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-neo4j>=0.9.0",
    "langchain-openai>=1.1.14",
    "langgraph>=1.1.8",
    "neo4j>=6.1.0",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=15.0.0",
    "tqdm>=4.67.3",
    "typer>=0.24.1",
]
